Dask Test Updates (#973)
* remove dask hackathon test

* reorganize and remove unnecessary tests

* remove dask worker files

* remove dask-tests-tmp directory

* remove dask_profiling.py

* update Makefile and MANIFEST.in
