+++ MANIFEST.in
@@ -2,6 +2,5 @@ include *.txt
 include LICENSE
 include README.md
 include featuretools/primitives/data/featuretools_unit_test_example.csv
-exclude featuretools/tests/entityset_tests/test_dask_hackathon.py
 recursive-exclude * __pycache__
 recursive-exclude * *.py[co]
